Reorganize file structure for button components
