+++ src/App.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { theme, GlobalStyles } from './global'
 import { ThemeProvider } from 'styled-components'
 import Logo from './components/Logo'
-import SearchButton from './components/SearchButton'
+import SearchButton from './components/buttons/SearchButton'
 import styled from 'styled-components'
 import { useDevice, isOneColumnLayout } from './util'
 import { Container } from './components/Container'
