+++ src/components/RecipeTab.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import RecipeCard from './RecipeCard'
-import { Button } from './Button'
+import { Button } from './buttons/Button'
 import { useDevice, isOneColumnLayout } from '../util'
 
 const TabContainer = styled.div`
