+++ src/components/buttons/Button.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { theme } from '../global'
+import { theme } from '../../global'
 
 const buttonColor = (variant) => variant === 'secondary' ? theme.accent : theme.text
 
@@ -21,21 +21,5 @@ export const Button = styled.button`
       0 0 0 3px ${theme.background},
       0 0 0 5px ${props => buttonColor(props.variant)};
     ;
-  }
-`
-
-export const IconButton = styled.button`
-  background: none;
-  border: none;
-  border-radius: 10px;
-  padding: 5px 0 0 5px;
-
-  :active {
-    filter: saturate(5);
-  }
-
-  :focus {
-    outline: none;
-    box-shadow: 0 0 0 2px ${theme.text};
   }
 `
